# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Games API",
        "version": "1.0.0",
        "description": "This is a Swagger UI for our Games API.\n\n### Error handling\nErrors use HTTP 400-599 with body:\n```\n{ \"error\": \"Cleartext explanation of the error that occurred\" }\n```"
    },
    "host": "localhost:8080",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
        "/games": {
            "get": {
                "summary": "List games (names only)",
                "description": "Returns an array of game names. Supports search `q` and ordering `order=asc|desc`.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "type": "string",
                        "description": "Case-insensitive substring search by name"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "default": "asc",
                        "description": "Sort by name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "Minecraft"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new game",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "name",
                                "price"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Hades"
                                },
                                "price": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 19.99
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/games/{id}": {
            "get": {
                "summary": "Get game details by id",
                "tags": [
                    "games"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete game by id",
                "tags": [
                    "games"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "definitions": {
                "Game": {
                    "type": "object",
                    "required": [
                        "id",
                        "name",
                        "price"
                    ],
                    "properties": {
                        "id": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "name": {
                            "type": "string"
                        },
                        "price": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                },
                "Error": {
                    "type": "object",
                    "required": [
                        "error"
                    ],
                    "properties": {
                        "error": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}